fixed build error in tools slug and added _component dividing tools page in two parts with client and wothout
